Refactoring WUBRG into smaller, more organized chunks.
